🛡️ Sentry: Add rigorous SIMD safety tests for vector scaling
- Updated `src/core/vector/sentry_tests.rs` with comprehensive tests for `scale_in_place`.
- Verified unaligned memory access safety (critical for SIMD) using a reusable helper.
- Verified `NaN`, `Infinity`, and `0.0` scalar handling.
- Verified large vector support (chunking + remainder).
- Added explicit coverage for AVX2/SSE2 paths via `unsafe` calls guarded by feature detection.
- Consolidated tests into `sentry_tests.rs` to adhere to DRY and avoid visibility issues.
